Split controller and Home Assistant client
